Renamed registration module to authentication
