+++ src-backend/authentication/tests/test_logout.py
